+++ rust/holo-public-api/src/providers/crud/mod.rs
@@ -1,0 +1,20 @@
+mod count;
+mod create;
+mod delete;
+mod delete_hard;
+mod find_one;
+mod get;
+mod get_owner;
+mod list;
+mod update;
+
+pub use count::count;
+pub use create::create;
+pub use delete::delete;
+#[allow(unused_imports)]
+pub use delete_hard::delete_hard;
+pub use find_one::find_one;
+pub use get::get;
+pub use get_owner::get_owner;
+pub use list::list;
+pub use update::update;
