+++ rust/holo-public-api/src/providers/mod.rs
@@ -1,8 +1,12 @@
 pub mod auth;
 pub mod config;
+#[allow(dead_code)]
 pub mod crud;
 pub mod docs;
 pub mod error_response;
 pub mod jwt;
 pub mod limiter;
 pub mod pagination;
+
+#[cfg(test)]
+mod tests;
